latest

+++ Inventory.Repository/Inventory.Repository/WarehouseRepository.cs
@@ -24,7 +24,7 @@ namespace Inventory.Repository
         }
 
         public static int WHaddressinsert(string WH_Name,string WH_ShortName,string contactperson,string jobposition,
-            int phone, int mobile,string Email,string Note,string bill_street,string bill_city,string bill_state,string bill_postalcode,
+            long phone, long mobile,string Email,string Note,string bill_street,string bill_city,string bill_state,string bill_postalcode,
             string bill_country, string ship_street, string ship_city, string ship_state, string ship_postalcode, string ship_country)
         {
             int count = SqlHelper.ExecuteNonQuery(ConnectionString, "insertwarehouse",WH_Name,WH_ShortName,contactperson, jobposition, phone,mobile,
