+++ Inventory/Models/Warehouse.cs
@@ -22,9 +22,9 @@ namespace Inventory.Models
 
         public string Job_position { get; set; }
         //[RegularExpression("([^[0-9]*$)")]
-        public int phone { get; set; }
+        public long phone { get; set; }
         //[RegularExpression("([^[0-9]*$)")]
-        public int Mobile { get; set; }
+        public long Mobile { get; set; }
         //[Required(ErrorMessage = "Please Enter Email Address")]
         //[Display(Name = "Email")]
         //[RegularExpression(@"^([a-zA-Z0-9_\.\-])+\@(([a-zA-Z0-9\-])+\.)+([a-zA-Z0-9]{2,4})+$",
